Veteriner butonu aktif edildi. Scrollbar eklendi. Code-cleaning (#72)
