+++ src/App.js
@@ -1,8 +1,7 @@
 import React from "react";
 import "./App.css";
-import MainViewContaier from "./components/MainViewContainer";
+import MainViewContaier from "./components/MainViewContainer/MainViewContainer";
 
 export default function App() {
- 
   return <MainViewContaier />;
 }
