+++ src/components/Footer/style.css
@@ -24,7 +24,7 @@
 }
 
 .ilce-item {
-  font-family: "SegoeUI";
+  font-family: SegoeUI, sans-serif;
   background-color: inherit;
   outline: none;
   border: none;
@@ -56,7 +56,7 @@
 }
 
 .city-item {
-  font-family: "SegoeUI";
+  font-family: SegoeUI, sans-serif;
   border: none;
   background-color: inherit;
   outline: hidden;
@@ -64,7 +64,7 @@
   font-size: 2.25rem;
   border: 1px white solid;
   border-radius: 10px;
-  padding: 15px 0;
+  padding: 1rem;
   flex-grow: 1;
   user-select: none;
   cursor: pointer;
@@ -156,7 +156,7 @@
     border: 1px white solid;
     border-radius: 10px;
     width: 110px;
-    padding: 5px 0;
+    padding: 0.325rem;
   }
   .city-item-active {
     background-color: white;
